UARG-122 moved enum method to it's own VariablType.java
